Correct code smell
Programmer : Camille 

Reviewer : Tessa

+++ src/main/java/gestion/metier/ActionSimple.java
@@ -43,7 +43,7 @@ public class ActionSimple extends Action {
     }
     
     @Override
-    public float valeur(Jour j) {
+    public double valeur(Jour j) {
         if(this.mapCours.containsKey(j) == true)
             return this.mapCours.get(j).getValeur();
         else
